+++ gce_ovf_import_tests/main.go
@@ -15,30 +15,15 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/xml"
-	"flag"
-	"fmt"
-	"io/ioutil"
 	"log"
-	"os"
-	"path/filepath"
 	"regexp"
-	"strings"
 	"sync"
 
-	"github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/junitxml"
-	"github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/test_config"
 	"github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/test_suites"
-)
-
-var (
-	testSuiteFilter = flag.String("test_suite_filter", "", "test suite filter")
-	testCaseFilter  = flag.String("test_case_filter", "", "test case filter")
-	outDir          = flag.String("out_dir", "/tmp", "junit xml directory")
-	testProjectID   = flag.String("test_project_id", "", "test project id")
-	testZone        = flag.String("test_zone", "", "test zone")
+	"github.com/GoogleCloudPlatform/compute-image-tools/test_common"
+	"github.com/GoogleCloudPlatform/compute-image-tools/test_common/junitxml"
+	"github.com/GoogleCloudPlatform/compute-image-tools/test_common/test_config"
 )
 
 var testFunctions = []func(context.Context, *sync.WaitGroup, chan *junitxml.TestSuite, *log.Logger, *regexp.Regexp, *regexp.Regexp, *testconfig.Project){
@@ -46,89 +31,5 @@ var testFunctions = []func(context.Context, *sync.WaitGroup, chan *junitxml.Test
 }
 
 func main() {
-	flag.Parse()
-	ctx := context.Background()
-
-	if len(strings.TrimSpace(*testProjectID)) == 0 {
-		fmt.Println("-test_project_id is invalid")
-		os.Exit(1)
-	}
-
-	if len(strings.TrimSpace(*testZone)) == 0 {
-		fmt.Println("-test_zone is invalid")
-		os.Exit(1)
-	}
-
-	pr := testconfig.GetProject(*testProjectID, *testZone)
-
-	var testSuiteRegex *regexp.Regexp
-	if *testSuiteFilter != "" {
-		var err error
-		testSuiteRegex, err = regexp.Compile(*testSuiteFilter)
-		if err != nil {
-			fmt.Println("-testSuiteFilter flag not valid:", err)
-			os.Exit(1)
-		}
-	}
-
-	var testCaseRegex *regexp.Regexp
-	if *testCaseFilter != "" {
-		var err error
-		testCaseRegex, err = regexp.Compile(*testCaseFilter)
-		if err != nil {
-			fmt.Println("-testCaseFilter flag not valid:", err)
-			os.Exit(1)
-		}
-	}
-
-	logger := log.New(os.Stdout, "[OvfImportTests] ", log.LstdFlags)
-	logger.Println("Starting...")
-
-	tests := make(chan *junitxml.TestSuite)
-	var wg sync.WaitGroup
-	for _, tf := range testFunctions {
-		wg.Add(1)
-		go tf(ctx, &wg, tests, logger, testSuiteRegex, testCaseRegex, pr)
-	}
-	go func() {
-		wg.Wait()
-		close(tests)
-	}()
-
-	var testSuites []*junitxml.TestSuite
-	for ret := range tests {
-		testSuites = append(testSuites, ret)
-		testSuiteOutPath := filepath.Join(*outDir, fmt.Sprintf("junit_%s.xml", ret.Name))
-		if err := os.MkdirAll(filepath.Dir(testSuiteOutPath), 0770); err != nil {
-			log.Fatal(err)
-		}
-
-		logger.Printf("Creating junit xml file: %s", testSuiteOutPath)
-		d, err := xml.MarshalIndent(ret, "  ", "   ")
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		if err := ioutil.WriteFile(testSuiteOutPath, d, 0644); err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	var buf bytes.Buffer
-	for _, ts := range testSuites {
-		if ts.Failures > 0 {
-			buf.WriteString(fmt.Sprintf("TestSuite %q has errors:\n", ts.Name))
-			for _, tc := range ts.TestCase {
-				if tc.Failure != nil {
-					buf.WriteString(fmt.Sprintf(" - %q: %s\n", tc.Name, tc.Failure.FailMessage))
-				}
-			}
-
-		}
-	}
-
-	if buf.Len() > 0 {
-		logger.Fatalf("%sExiting with exit code 1", buf.String())
-	}
-	logger.Print("All test cases completed successfully.")
+	testcommon.LaunchTests(testFunctions, "[OvfImportTests]")
 }
