+++ test-infra/prowjobs/gce-image-import-export-tests/Dockerfile
@@ -18,10 +18,10 @@ RUN apk add --no-cache git
 # Build test runner
 WORKDIR /
 RUN go get -d github.com/GoogleCloudPlatform/compute-image-tools/gce_image_import_export_tests
-RUN CGO_ENABLED=0 go build -o /gce_ovf_import_test_runner github.com/GoogleCloudPlatform/compute-image-tools/gce_image_import_export_tests
-RUN chmod +x /gce_ovf_import_test_runner
+RUN CGO_ENABLED=0 go build -o /gce_image_import_export_test_runner github.com/GoogleCloudPlatform/compute-image-tools/gce_image_import_export_tests
+RUN chmod +x /gce_image_import_export_test_runner
 
-# Build binaries to test by test runner
+# Build binaries to test
 WORKDIR /
 RUN go get -d github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/gce_vm_image_import
 RUN CGO_ENABLED=0 go build -o /gce_vm_image_import github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/gce_vm_image_import
@@ -34,12 +34,7 @@ RUN chmod +x /gce_vm_image_export
 FROM gcr.io/compute-image-tools-test/wrapper:latest
 ENV GOOGLE_APPLICATION_CREDENTIALS /etc/compute-image-tools-test-service-account/creds.json
 COPY --from=0 /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
-COPY --from=0 /gce_ovf_import_test_runner gce_ovf_import_test_runner
+COPY --from=0 /gce_image_import_export_test_runner gce_image_import_export_test_runner
 COPY --from=0 /gce_vm_image_import gce_vm_image_import
 COPY --from=0 /gce_vm_image_export gce_vm_image_export
-# ??
-COPY /gce_image_import_export_tests/scripts/ /gce_image_import_export_tests/scripts/
-# ??
-COPY /daisy_integration_tests/scripts/ /daisy_integration_tests/scripts/
-COPY /daisy_workflows/ /daisy_workflows/
 ENTRYPOINT ["./wrapper", "./gce_image_import_export_test_runner"]
