+++ test_common/test_config/project_config_test.go
@@ -25,7 +25,7 @@ func TestGetZone(t *testing.T) {
 
 	got := map[string]int{"zone1": 0, "zone2": 0, "zone3": 0}
 	for i := 0; i < 60; i++ {
-		zone := p.GetZone()
+		zone := GetZone()
 		for z := range got {
 			if z == zone {
 				got[z]++
